+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/lang/file/LibraryProtobufFileResolver.kt
@@ -19,4 +19,25 @@ class LibraryProtobufFileResolver : FileResolver {
                 it.findFileByRelativePath(path)?.takeIf { it.exists() }
             }
     }
+
+    override fun collectProtobuf(path: String, project: Project): Iterable<VirtualFile> {
+        val result = mutableListOf<VirtualFile>()
+        ProjectRootManager.getInstance(project).orderEntries().allSourceRoots.forEach {
+            val directory = it.findFileByRelativePath(path) ?: return@forEach
+            if (!directory.isDirectory) return@forEach
+            FileResolver.collectProtobuf(directory, result)
+        }
+        return result
+    }
+
+    override fun collectProtobuf(path: String, module: Module): Iterable<VirtualFile> {
+        val result = mutableListOf<VirtualFile>()
+        ModuleRootManager.getInstance(module).orderEntries()
+            .withoutModuleSourceEntries().allSourceRoots.forEach {
+                val directory = it.findFileByRelativePath(path) ?: return@forEach
+                if (!directory.isDirectory) return@forEach
+                FileResolver.collectProtobuf(directory, result)
+            }
+        return result
+    }
 }