+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/lang/file/ModuleSourceProtobufFileResolver.kt
@@ -15,4 +15,14 @@ class ModuleSourceProtobufFileResolver : FileResolver {
             it.findFileByRelativePath(path)?.takeIf { it.exists() }
         }
     }
+
+    override fun collectProtobuf(path: String, module: Module): Iterable<VirtualFile> {
+        val result = mutableListOf<VirtualFile>()
+        ModuleRootManager.getInstance(module).sourceRoots.forEach {
+            val directory = it.findFileByRelativePath(path) ?: return@forEach
+            if (!directory.isDirectory) return@forEach
+            FileResolver.collectProtobuf(directory, result)
+        }
+        return result
+    }
 }