+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/lang/psi/primitive/element/ProtobufPackageName.kt
@@ -101,7 +101,7 @@ interface ProtobufPackageName :
         val name = name() ?: return null
         return LookupElementBuilder.create(name)
             .withTypeText("package")
-            .withLookupString(name)
+            .withPresentableText(name)
             .withIcon(getIcon(false))
             .withPsiElement(this)
     }