+++ CHANGELOG.md
@@ -3,8 +3,14 @@
 # IntelliJ Platform Plugin Template Changelog
 
 ## [Unreleased]
+### Added
+- Duplicate and unused import annotator
+- Import optimizer
+- Auto-completion for import
 ### Changed
 - Fix some keywords highlight in typename(etc. 'rpc')
+- Refactor psi structure by default implementation
+- New reference and resolving cache implementation
 
 ## [0.0.3]
 ### Added